+++ src/Mews.Fiscalization.Core.Tests/Model/TaxpayerIdentificationNumberTests.cs
@@ -71,24 +71,5 @@ namespace Mews.Fiscalization.Core.Tests.Model
             Assert.Throws<ArgumentNullException>(() => new TaxpayerIdentificationNumber(null, "ABC1234567"));
             Assert.Throws<ArgumentNullException>(() => new TaxpayerIdentificationNumber(country, null));
         }
-
-        [Test]
-        public void CreatingCountryWithInvalidCountryCodeFails()
-        {
-            Assert.Throws<ArgumentException>(() => new EuropeanCountry("US"));
-            Assert.Throws<ArgumentException>(() => new Country(""));
-            Assert.Throws<ArgumentNullException>(() => new Country(null));
-        }
-
-        [Test]
-        public void CreatingCountryWithValidCountryCodeSucceeds()
-        {
-            Assert.DoesNotThrow(() => new EuropeanCountry("CZ"));
-            Assert.DoesNotThrow(() => new Country("CZ"));
-            Assert.DoesNotThrow(() => new Country("US"));
-            Assert.IsTrue(EuropeanCountry.IsValid("CZ"));
-            Assert.IsTrue(Country.IsValid("CZ"));
-            Assert.IsTrue(Country.IsValid("US"));
-        }
     }
 }
